requirements.txt
fastapi
uvicorn
PyPDF2
chromadb
python-dotenv
openai